+++ app/Species.php
@@ -2,10 +2,13 @@
 
 namespace App;
 
+use App\Traits\Species\Icons;
 use Illuminate\Database\Eloquent\Model;
 
 class Species extends Model
 {
+    use Icons;
+    
     const HARDINESS_CA = ['0', '0a', '0b', '1', '1a', '1b', '2', '2a', '2b', '3', '3a', '3b', '4', '4a', '4b', '5', '5a', '5b', '6', '6a', '6b', '7', '7a', '7b', '8', '8a', '8b', '9', '9a', '9b'];
     const SUN = ['full', 'partial', 'shade'];
     const WATER = [1, 2, 3, 4, 5];