+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # NEXT
 
 - Can add additional informations to _species_: uses, ornemental, horticulture.
+- Page to show species informations.
 
 # 0.2.0 - Pre-release (2019-08-24)